refactor: moved `OnboardingPrivacyAdvice`
